requirements.txt
flask==2.0.1
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.47.0
openai==0.27.0
pymongo==4.1.1
python-dotenv==0.19.0
requests==2.26.0
dnspython==2.2.1
flask-cors==3.0.10
gunicorn==20.1.0
python-dateutil==2.8.2 